refactor: split production CLI root
